Add latest version of `feature-car`

+++ src/application/services/Car/DeleteCarService.ts
@@ -1,18 +1,20 @@
 import { CarsRepository } from '../../../domain/repositories/CarsRepository';
+import { OrderRepository } from '../../../domain/repositories/OrderRepository';
 
 export class DeleteCarService {
-  private CarsRepository: CarsRepository;
-
-  constructor(CarsRepository: CarsRepository) {
-    this.CarsRepository = CarsRepository;
-  }
+  constructor(
+    private CarsRepository: CarsRepository,
+    private OrderRepository: OrderRepository
+  ) { }
 
   public async execute(id: string): Promise<void> {
-    // TODO: Um carro não deve ser excluído se estiver em algum pedido aberto
     const car = await this.CarsRepository.findById(id);
 
     if (!car)
       throw new Error("Carro não encontrado.");
+
+    if (await this.OrderRepository.findOpenOrderByCarId(id))
+      throw new Error("O carro não pode ser excluído pois tem pedidos em aberto.");
 
     await this.CarsRepository.remove(car);
   }
